[swss] Suppress error messages when depend containers are not running.
DPU SONiC runs a subset of features. Some of the features are disabled on
compilation time. When the feature is disabled on compilation time the container
for the feature is not created on system start. Swss depends on teamd feature
and monitors its state. When teamd container doesn't exist "docker inspect"
command triggered by swss.sh script returns an error message:
sonic INFO swss.sh[5535]: Error: No such object: teamd
This error is expected and shouldn't be logged.

The same applies to docker-wait-any script. The error should not be logged and
script should continue monitoring of the container even if it doesn't exist initially.