lldpmgrd: Interrupt select on signal
When trying to stop the lldp docker, the SIGTERM sinal is sent to
lldpmgrd, but the handling of this signal is delayed because most of the
time, lldpmgrd is going to be on the sel.select line, which goes to
swss-common's C++ code where it calls `select()`. When the SIGTERM
arrives, the `select()` calls returns early with EINTR, but because
`interrupt_on_signal` is false, swss-common goes back into the
`select()` call.

Because the SIGTERM handler is in Python, but the current call is in
C++, the SIGTERM handling doesn't take effect until the call stack
returns to Python code, which will take 10 seconds. By that time, docker
will kill the docker with SIGKILL.

To avoid that, set `interrupt_on_signal` to true, which tells
swss-common to return to Python code upon a signal, which will let any
signal handlers run quickly.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>